pyproject.toml
[project]
name = "llm-meditation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "accelerate",
    "sae-lens>=4.0",
    "nnsight>=0.3",
    "huggingface-hub",
    "pyyaml",
    "pandas",
    "numpy",
    "scipy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "tqdm",
    "datasets",
    "requests",
]

[project.optional-dependencies]
dev = ["pytest", "ipython", "jupyter", "ruff"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
